requirements.txt
farm-haystack[inference]
farm-haystack[faiss]
datasets~=2.14.6
openai~=1.3.4
ragas~=0.0.22
nltk~=3.8.1
langchain~=0.0.339
llama-index~=0.9.4
xmltodict~=0.13.0
backoff~=2.2.1
matplotlib~=3.8.2
bert-score~=0.3.13
evaluate~=0.4.1
sentence-transformers~=2.2.2